+++ src/lightning_habana/fabric/__init__.py
@@ -15,12 +15,14 @@
 
 from lightning_habana.fabric.accelerator import HPUAccelerator
 from lightning_habana.fabric.plugins.io_plugin import HPUCheckpointIO
+from lightning_habana.fabric.strategies.ddp import HPUDDPStrategy
 from lightning_habana.fabric.plugins.precision import HPUPrecision
 from lightning_habana.fabric.strategies.parallel import HPUParallelStrategy
 from lightning_habana.fabric.strategies.single import SingleHPUStrategy
 
 __all__ = [
     "HPUAccelerator",
+    "HPUDDPStrategy",
     "HPUParallelStrategy",
     "SingleHPUStrategy",
     "HPUPrecision",